backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
SQLAlchemy==2.0.23
psycopg2==2.9.9
python-dotenv==1.0.0

# pip install -r requirements.txt